Compartmentalized the different parts of the BarContainer.
