+++ style/style.css
@@ -52,10 +52,6 @@
     margin-left: 5em;
   }
 
-  .shift-right {
-    right: 5%;
-  }
-
   .checkbox {
       width: 1.75em;
       height: 1.75em;
@@ -66,10 +62,12 @@
     display: none;
     visibility: hidden;
   }
-
-
+  .shift-right {
+    right: 5%;
+  }
 }
-/* This is for the plus buttom for inserting todo*/
+
+/* This is for the plus buttom for creating new todo*/
 
 .PlusBtn {
   position: absolute;
